requirements.txt
google-genai
python-dotenv
yfinance
pandas
numpy
matplotlib